Add coarse stats about critical strikes to detailed breakdown
These really aren't included elsewhere, because critical strikes are
relative to the other person's ttl, making this an ODE in many different
variables.  Instead I just compute an "average" critical strike rate to
get an idea of its efficacy for changing things.  I expected it to be low,
but the results seemed to suggest it's a much lower factor than I even
envisioned.