Add a quit-strategy option, showing how much quitting would affect everyone
I have often wondered whether it would be beneficial when I am in a quest
for me to take a talking penalty in order to penalize everyone when I have
lots of TTL left but someone else who is at a higher level is about to
level again.  The thing is, I'd get the talk penalty plus the p15, and I'd
lose out on the 25% TTL reduction from completing the quest, whereas others
would only get the p15.  But a p15 is worse for higher level folks and any
penalty is worse for those about to reach the next level than those who are
far away.  Since I have good models for how much time someone has until the
next level based on their current TTL and itemsums and alignment, take the
guesswork out by providing a function which will show how much longer it is
expected that everyone will take to get to the next level if someone quits
a quest.

Similarly, the question can be asked when I'm not in a quest but others are
whether getting one of those people to quit is better for me or not.
Usually the answer would be yes (keep them from getting a 25% TTL
reduction, plus the p15 would hurt others ahead of me more), but perhaps not
if I am near the next level.