Add ability to show zero ttl-slope point
Given "anti-burn" and that penalties are exponential in the level,
it stands to reason that the continuous equation model has a point at
which the ttl-burndown slope hits 0.  For low-level characters, clearly
that point will be well below ttl=0, but for high-level characters it
may be at some ttl above 0.  Calculate these values.