Be more aggressive in marking unknown-offline folks as offline
Folks are very unlikely to remain online for several days without idlerpg
selecting them for something (battles, quests, godsends, calamities, etc.),
so just mark them offline if they've been unknown-offline for longer than
the optimistic-time-to-level; don't bother waiting for the
full-time-to-level.  I knew for more than a month that dlawler was offline
when the old logic continued claiming it didn't know he was offline.