Switch primary logs over to new irssi ones
Disable additional logs from Atychiphobe for now, since they go back long
before the new irssi logs for elijah -- having such long logs causes
problems with the levelling stuff since it sets state when the character
gets new items at new levels which is expected to be unset by a matching
line from the main log.  That matching line isn't there in the main log
simply because it doesn't go that far back in history and we get a KeyError
for 'ignore_level'.  I'll need to fix that up later.