guess_seed.py: Fix some off-by-one errors
I don't think I actually triggered these anywhere since 2**48 is a
really big number and we tend to divide it into hundreds or
thousands of equal-sized segments and the values at the end are
unlikely to be the exact one we are looking for.  But, better to be
correct and include them.