Fix a corner case issue with item transfer and item godsend
j lost his weapon in a fierce battle, followed shortly by getting an
item godsend on his charm.  Being evil, there's a .9 factor on all his
itemchanges, and with no battle between these events that he was involved
in, it threw off my calculations a little bit and resulted the recording of
a level 88 charm for him instead of level 100.  This error went undetected
until he lost his charm in a fierce battle, which resulted in hitting the
"Mistmatch for j; 88 vs 100" error message.