levelling.py: Handle quests a bit more robustly
If a quit in the middle of a quest, I won't see the "elijah's prudence
and self regard..." message about quitting, so I don't record the quest as
having been aborted.  If the next quest starting message comes and it's of
a different type, make sure to clear the quest_time_left so that code that
keys off of that doesn't assume the wrong type of quest is in action.