levelling.py: Improve offline handling
We wanted to mark people as offline from whom we clearly hadn't seen
activity in a long enough time.  Problem was, if they had a really low
itemsum, our "long enough time" measurement involved the amount of time
they'd take to reach the next level, and a low enough itemsum means they
have an infinite amount of time necessary to reach the next level.  So,
instead, take the minimum of the recorded time to next level and the
projected time needed to reach the next level and use that as approximate
time left in this particular calculation.